ajoute le fonctionalite de emprente et modifier le regester et partie admin

+++ bibiliotheque/admin_system/views.py
@@ -7,64 +7,86 @@ from django.core.paginator import Paginator
 
 # Create your views here.
 def ajout_liver(request):
-    if request.method=='GET':
-        return render(request,'admin/ajoute_liver.html')
+    if 'admin' in request.session:
+        if request.method=='GET':
+            return render(request,'admin/ajoute_liver.html')
+        else:
+                
+            titre=request.POST['titre']
+            description=request.POST['description']
+            auteur=request.POST['auteur']
+            date_publication=request.POST['date_publication']
+            genre=request.POST['genre']
+            couverture=request.FILES['couverture']
+            liver=Livre(titre=titre,description=description,auteur=auteur,date_publication=date_publication,genre=genre,couverture=couverture)
+            liver.save()
+            return HttpResponseRedirect(reverse('affiche_livre'))
     else:
-            
-        titre=request.POST['titre']
-        description=request.POST['description']
-        auteur=request.POST['auteur']
-        date_publication=request.POST['date_publication']
-        genre=request.POST['genre']
-        couverture=request.FILES['couverture']
-        liver=Livre(titre=titre,description=description,auteur=auteur,date_publication=date_publication,genre=genre,couverture=couverture)
-        liver.save()
-        return HttpResponseRedirect(reverse('affiche_livre'))
+         return HttpResponseRedirect('/login')
 def update_liver(request,id):
-    if request.method=='GET':
-        liver=Livre.objects.filter(id=id).first()
-        return render(request,'admin/update.html',{'livre':liver})
+    if 'admin' in request.session:
+        if request.method=='GET':
+            liver=Livre.objects.filter(id=id).first()
+            return render(request,'admin/update.html',{'livre':liver})
+        else:
+            liver=Livre.objects.filter(id=id).first()
+            
+            liver.titre=request.POST['titre']
+            liver.description=request.POST['description']
+            liver.auteur=request.POST['auteur']
+            liver.date_publication=request.POST['date_publication']
+            liver.genre=request.POST['genre']
+            # liver['couverture']=request.POST['couverture']
+            liver.save()
+            return HttpResponseRedirect(reverse('affiche_livre'))
     else:
-        liver=Livre.objects.filter(id=id).first()
-        
-        liver.titre=request.POST['titre']
-        liver.description=request.POST['description']
-        liver.auteur=request.POST['auteur']
-        liver.date_publication=request.POST['date_publication']
-        liver.genre=request.POST['genre']
-        # liver['couverture']=request.POST['couverture']
-        liver.save()
-        return HttpResponseRedirect(reverse('affiche_livre'))
+        return HttpResponseRedirect('/login')
 def update_client(request,id):
-    if request.method=='GET':
-        client=Clients.objects.filter(id=id).first()
-        return render(request,'admin/update_cliens.html',{'client':client})
+    if 'admin' in request.session:
+        if request.method=='GET':
+            client=Clients.objects.filter(id=id).first()
+            return render(request,'admin/update_cliens.html',{'client':client})
+        else:
+            client=Clients.objects.filter(id=id).first()
+            
+            client.nom=request.POST['nom']
+            client.prenom=request.POST['prenom']
+            client.email=request.POST['email']
+            client.tel=request.POST['tel']
+            client.save()
+            return HttpResponseRedirect(reverse('affiche_cliens'))
     else:
-        client=Clients.objects.filter(id=id).first()
-        
-        client.nom=request.POST['nom']
-        client.prenom=request.POST['prenom']
-        client.email=request.POST['email']
-        client.tel=request.POST['tel']
-        client.save()
-        return HttpResponseRedirect(reverse('affiche_cliens'))
+        return HttpResponseRedirect('/login')
 def affiche_livre(request):
-    livres=Livre.objects.all()
-    paginator = Paginator(livres, 4)
-    
-    page_number = request.GET.get('page')
-    page_obj = paginator.get_page(page_number)
-    return render(request,'admin/affiche_livre.html',{'livres':page_obj})
+    if 'admin' in request.session:
+        email_admin=request.session['admin']
+        livres=Livre.objects.all()
+        paginator = Paginator(livres, 4)
+        
+        page_number = request.GET.get('page')
+        page_obj = paginator.get_page(page_number)
+        return render(request,'admin/affiche_livre.html',{'livres':page_obj,'admin_conect':email_admin})
+    else:
+         return HttpResponseRedirect('/login')
 def affiche_cliens(request):
-    clients=Clients.objects.all()
-    return render(request,'admin/affiche_cliens.html',{'clients':clients})
+    if 'admin' in request.session:
+        email_admin=request.session['admin']
+        clients=Clients.objects.all()
+        return render(request,'admin/affiche_cliens.html',{'clients':clients,'admin_conect':email_admin})
+    else:
+        return HttpResponseRedirect('/login')
 def delete(request,id):
-    liver=Livre.objects.filter(id=id).first()
-    liver.delete()
-    return HttpResponseRedirect(reverse('affiche_livre'))
+    if 'admin' in request.session:
+        liver=Livre.objects.filter(id=id).first()
+        liver.delete()
+        return HttpResponseRedirect(reverse('affiche_livre'))
+    else:
+        return HttpResponseRedirect('/login')
 def test(request):
     return render(request,'admin/test.html')
 def ajoute_clients(request):
+    if 'admin' in request.session:
+        
         nom=request.POST['nom']
         prenom=request.POST['prenom']
         email=request.POST['email']
@@ -74,9 +96,20 @@ def ajoute_clients(request):
         user=Clients(nom=nom,prenom=prenom,email=email,password=password,tel=tel,age=age)
         user.save()
         return HttpResponseRedirect(reverse('affiche_cliens'))
+    else:
+        return HttpResponseRedirect('/login')
 def delete_clients(request,id):
-    client=Clients.objects.filter(id=id).first()
-    client.delete()
-    return HttpResponseRedirect(reverse('affiche_cliens'))
+    if 'admin' in request.session:
+        
+        client=Clients.objects.filter(id=id).first()
+        client.delete()
+        return HttpResponseRedirect(reverse('affiche_cliens'))
+    else:
+        return HttpResponseRedirect('/login')
 def login(request):
     return HttpResponseRedirect('/login')
+def deconection(request):
+    request.session.flush()
+    return HttpResponseRedirect('/login')
+def affiche_liver_emprente(request):
+    pass
